requirements.txt
streamlit
openai
python-dotenv
pdfplumber
python-docx
pyyaml
anthropic
httpx
fastapi
uvicorn
python-multipart
streamlit
